feat: support for setting ALB TLS Policy
Add a variable to set the ALB Ingress TLS policy. If not set the default value will be used.
https://kubernetes-sigs.github.io/aws-load-balancer-controller/v2.4/guide/ingress/annotations/#ssl-policy

Note that the alb.ingress.kubernetes.io/ssl-policy annotation must be the same for _all_ ingresses in the
cluster. This feature will create a default ingress with the wanted policy to avoid updating all ingressess
when changing the policy