feat: k8s-cilium-kubeproxy-update (#200)
* feat(k8s): add kube_proxy configuration and enable node port

Add a kube_proxy block to manage the proxy settings based on the  
presence of CNI configurations. Enable node port functionality to  
support external access to services. These changes improve the  
flexibility and usability of the Kubernetes deployment.

* feat(kubernetes): add Cilium patch role for RBAC permissions

Introduce a new role and role binding for Cilium to manage secrets 
within the kube-system namespace. This update creates necessary 
permissions for the Cilium service account, ensuring it can access 
secrets as needed. The changes enhance the security posture and 
functionality of Cilium by defining clear RBAC rules. Additionally, 
the new role is integrated into the addons configuration to ensure 
it is applied during deployment.